+++ opensrp-reveal/src/main/java/org/smartregister/reveal/contract/BaseContract.java
@@ -1,0 +1,26 @@
+package org.smartregister.reveal.contract;
+
+import android.support.annotation.NonNull;
+
+import com.mapbox.geojson.Feature;
+
+import org.json.JSONArray;
+import org.smartregister.domain.Task.TaskStatus;
+
+/**
+ * Created by samuelgithengi on 3/25/19.
+ */
+public class BaseContract {
+
+    public interface BasePresenter {
+
+        void onSprayFormSaved(@NonNull String structureId, @NonNull String taskIdentifier,
+                              @NonNull TaskStatus taskStatus, @NonNull String businessStatus);
+
+        void onStructureAdded(Feature feature, JSONArray featureCoordinates);
+
+        void onFormSaveFailure(String eventType);
+
+        void onMosquitoCollectionFormSaved();
+    }
+}
