+++ opensrp-reveal/src/main/java/org/smartregister/reveal/interactor/TaskRegisterFragmentInteractor.java
@@ -8,7 +8,6 @@ import net.sqlcipher.database.SQLiteDatabase;
 
 import org.smartregister.cursoradapter.SmartRegisterQueryBuilder;
 import org.smartregister.family.util.DBConstants;
-import org.smartregister.repository.StructureRepository;
 import org.smartregister.reveal.application.RevealApplication;
 import org.smartregister.reveal.contract.TaskRegisterFragmentContract;
 import org.smartregister.reveal.model.TaskDetails;
