+++ opensrp-reveal/src/main/java/org/smartregister/reveal/interactor/TaskRegisterInteractor.java
@@ -3,6 +3,7 @@ package org.smartregister.reveal.interactor;
 import org.smartregister.configurableviews.ConfigurableViewsLibrary;
 import org.smartregister.configurableviews.helper.ConfigurableViewsHelper;
 import org.smartregister.reveal.application.RevealApplication;
+import org.smartregister.reveal.contract.BaseContract;
 import org.smartregister.reveal.contract.TaskRegisterContract;
 import org.smartregister.reveal.util.AppExecutors;
 
@@ -11,13 +12,14 @@ import java.util.List;
 /**
  * Created by samuelgithengi on 3/14/19.
  */
-public class TaskRegisterInteractor implements TaskRegisterContract.Interactor {
+public class TaskRegisterInteractor extends BaseInteractor implements TaskRegisterContract.Interactor {
 
     private ConfigurableViewsHelper viewsHelper;
 
     private AppExecutors appExecutors;
 
-    public TaskRegisterInteractor() {
+    public TaskRegisterInteractor(BaseContract.BasePresenter presenterCallBack) {
+        super(presenterCallBack);
         viewsHelper = ConfigurableViewsLibrary.getInstance().getConfigurableViewsHelper();
         appExecutors = RevealApplication.getInstance().getAppExecutors();
     }
